qiao update

+++ src/main/java/com/hongshen/sran_service/controller/AlarmLibController.java
@@ -29,19 +29,33 @@ public class AlarmLibController extends BaseController{
                                    @PathParam("generation")String generation,
                                    @HeaderParam("Auth-Token")String authToken){
 
+        String msg = "";
+        List<JSONObject> alarmList = null;
         JSONObject result = new JSONObject();
         NetObjBase obj = objFactory.getNetObj(supplier, generation);
-        List<JSONObject> alarmList = obj.getAlarmLibService().getAlarmList();
+        if (obj == null) {
+            msg += "Supplier or Generation is null.";
 
-        if (alarmList == null || alarmList.isEmpty()){
-            result.put("result", Constants.FAIL);
-            result.put("msg", Constants.MSG_NO_DATA);
+        }else{
+            alarmList = obj.getAlarmLibService().getAlarmList();
 
-        } else {
+            if (alarmList == null || alarmList.isEmpty()) {
+
+                msg += "Get alarmList failed.";
+            }
+        }
+        if (msg.length() == 0){
+
             result.put("result", Constants.SUCCESS);
             result.put("data", alarmList);
-        }
 
+        }else{
+
+            result.put("result", Constants.FAIL);
+
+            result.put("msg", Constants.MSG_NO_DATA + msg);
+
+        }
         return result;
     }
 
@@ -54,19 +68,37 @@ public class AlarmLibController extends BaseController{
                                    @PathParam("alarmName")String alarmName,
                                    @HeaderParam("Auth-Token")String authToken){
 
+        String msg ="";
+        JSONObject alarm = null;
         JSONObject result = new JSONObject();
         NetObjBase obj = objFactory.getNetObj(supplier,generation);
-        JSONObject alarm = obj.getAlarmLibService().getAlarmByName(alarmName);
+        if(obj == null) {
+            msg +="Supplier or Generation is null.";
+        }else {
+            if(alarmName == null) {
+                msg +="Alarm is null";
+            }else {
+                alarm = obj.getAlarmLibService().getAlarmByName(alarmName);
 
-        if (alarm == null || alarm.isEmpty()){
-            result.put("result", Constants.FAIL);
-            result.put("msg", Constants.MSG_NO_DATA);
+                if (alarm == null || alarm.isEmpty()) {
+//                result.put("result", Constants.FAIL);
+//                result.put("msg", Constants.MSG_NO_DATA);
+                    msg += "Alarm is null.";
+                }
+            }
+        }
+        if (msg.length() == 0){
 
-        } else {
             result.put("result", Constants.SUCCESS);
             result.put("data", alarm);
-        }
 
+        }else{
+
+            result.put("result", Constants.FAIL);
+
+            result.put("msg", Constants.MSG_NO_DATA + msg);
+
+        }
         return result;
     }
 
@@ -79,18 +111,41 @@ public class AlarmLibController extends BaseController{
                                       @PathParam("alarmName")String alarmName,
                                       @HeaderParam("Auth-Token")String authToken,
                                       JSONObject param){
-
+        String msg ="";
         JSONObject result = new JSONObject();
         NetObjBase obj = objFactory.getNetObj(supplier,generation);
-        int i = obj.getAlarmLibService().updateAlarmByName(alarmName,param);
+        if (obj == null){
+            msg +="Supplier or Generation is null.";
+        }else {
+            if (alarmName ==null){
+                msg +="AlarmName is null.";
+            }else {
+                if(param == null){
+                    msg += "Param is null.";
+                }else {
+                    try {
+                        int i = obj.getAlarmLibService().updateAlarmByName(alarmName, param);
+                        if (i > 0) {
+                            msg +="Update failed.";
 
-        if (i > 0){
+                        } else {
+                            result.put("result", Constants.FAIL);
+                            result.put("msg", Constants.MSG_UPDATE_FAILED);
+                        }
+                    }catch (Exception e){
+                        msg += "Parameters is Error.";
+                    }
+                }
+            }
+        }
+        if (msg.length() == 0){
+
             result.put("result", Constants.SUCCESS);
-            result.put("msg", Constants.MSG_UPDATE_OK);
+            result.put("data", Constants.MSG_UPDATE_OK);
 
-        } else {
+        }else{
             result.put("result", Constants.FAIL);
-            result.put("msg", Constants.MSG_UPDATE_FAILED);
+            result.put("msg", Constants.MSG_UPDATE_FAILED + msg);
 
         }
         return result;
@@ -108,17 +163,28 @@ public class AlarmLibController extends BaseController{
         JSONObject result = new JSONObject();
         String msg = "";
         NetObjBase obj = objFactory.getNetObj(supplier,generation);
-
-        if (param.getString("alarmNameId") != ""){
-            int i = obj.getAlarmLibService().addAlarmIndex(param);
-            if (i == 0){
-                msg = "alarmNameId is exist.";
-            }
-        }
-        if (msg.length() == 0 && param.getString("alarmName") != "") {
-            int i = obj.getAlarmLibService().addAlarm(param);
-            if (i == 0){
-                msg = "alarmName is exist.";
+        if(obj == null){
+            msg +="Supplier or Generation is null.";
+        }else {
+            if (param == null){
+                msg += "Param is null.";
+            }else {
+                try {
+                    if (param.getString("alarmNameId") != "") {
+                        int i = obj.getAlarmLibService().addAlarmIndex(param);
+                        if (i == 0) {
+                            msg = "alarmNameId is exist.";
+                        }
+                    }
+                    if (msg.length() == 0 && param.getString("alarmName") != "") {
+                        int i = obj.getAlarmLibService().addAlarm(param);
+                        if (i == 0) {
+                            msg = "alarmName is exist.";
+                        }
+                    }
+                }catch (Exception e){
+                    msg += "Parameters is Error.";
+                }
             }
         }
 
@@ -128,7 +194,7 @@ public class AlarmLibController extends BaseController{
 
         } else {
             result.put("result", Constants.FAIL);
-            result.put("msg", msg);
+            result.put("msg", Constants.MSG_ADD_FAILED + msg);
 
         }
 
@@ -145,17 +211,27 @@ public class AlarmLibController extends BaseController{
                                    @HeaderParam("Auth-Token")String authToken) {
 
         JSONObject result = new JSONObject();
+        String msg = "";
         NetObjBase obj = objFactory.getNetObj(supplier,generation);
-        int ret = obj.getAlarmLibService().deleteAlarmByName(alarmName);
+        if (obj == null){
+            msg +="Supplier or Generation is null.";
+        }else {
+            try {
+                int ret = obj.getAlarmLibService().deleteAlarmByName(alarmName);
 
-        if (ret > 0){
+                if (ret <= 0) {
+                    msg +="deleteAlarm is Failed";
+                }
+            }catch (Exception e) {
+                msg += "Parameters is Error.";
+            }
+        }
+        if (msg.length() == 0){
             result.put("result", Constants.SUCCESS);
             result.put("msg", Constants.MSG_DELETE_OK);
-
         } else {
             result.put("result", Constants.FAIL);
-            result.put("msg", Constants.MSG_DELETE_FAILED);
-
+            result.put("msg", Constants.MSG_DELETE_FAILED + msg);
         }
 
         return result;
